rework predicate; remove deprecated; fix D64 cases

+++ SYCL_ESIMD_EMBARGO/ESIMD/EMBARGO/lsc/lsc_surf_load_u64.cpp
@@ -24,19 +24,21 @@ int main(void) {
   bool passed = true;
 
   // non transpose
-  passed &= test<0, uint32_t, 1, 4, 32, 1, false>(rand());
-  passed &= test<1, uint32_t, 1, 4, 32, 2, false>(rand());
-  passed &= test<2, uint32_t, 1, 4, 16, 2, false>(rand());
-  passed &= test<3, uint32_t, 1, 4, 4, 1, false>(rand());
-  passed &= test<4, uint32_t, 1, 1, 1, 1, false>(1);
-  passed &= test<5, uint32_t, 2, 1, 1, 1, false>(1);
-  // passed &= test<6, uint32_t, 1, 4, 8, 2>(rand()); // merge fail
-  // passed &= test<7, uint32_t, 1, 4, 8, 3>(rand()); // exec fail
+  passed &= test<0, uint64_t, 1, 4, 32, 1, false>(rand());
+  passed &= test<1, uint64_t, 1, 4, 32, 2, false>(rand());
+  passed &= test<2, uint64_t, 1, 4, 16, 2, false>(rand());
+  passed &= test<3, uint64_t, 1, 4, 4, 1, false>(rand());
+  passed &= test<4, uint64_t, 1, 1, 1, 1, false>(1);
+  passed &= test<5, uint64_t, 2, 1, 1, 1, false>(1);
+
+  // IGC prohibits exec_size less than simd_size when vector size > 1
+  // passed &= test<6, uint64_t, 1, 4, 8, 2, false>(rand());
+  // passed &= test<7, uint64_t, 1, 4, 8, 3, false>(rand());
 
   // transpose
-  passed &= test<8, uint32_t, 1, 4, 1, 32, true>();
-  passed &= test<9, uint32_t, 2, 2, 1, 16, true>();
-  passed &= test<10, uint32_t, 4, 4, 1, 4, true>();
+  passed &= test<8, uint64_t, 1, 4, 1, 32, true>();
+  passed &= test<9, uint64_t, 2, 2, 1, 16, true>();
+  passed &= test<10, uint64_t, 4, 4, 1, 4, true>();
 
   std::cout << (passed ? "Passed\n" : "FAILED\n");
   return passed ? 0 : 1;
